Move examples out of src
