+++ example/index.js
@@ -1,7 +1,10 @@
+/* global document */
+
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import './reset.css';
+import './app.css';
 
 const appRoot = document.createElement('div');
